refactor(canvas): extract ExcludedPathsField, fix hardcoded enums and error handling
