+++ apps/app/src/pages/CanvasPage/CanvasNodePropertiesPanel/index.ts
@@ -1,1 +1,2 @@
 export * from "./CanvasNodePropertiesPanel";
+export * from "./ExcludedPathsField";
